Restructure Code for C++
