+++ projectq/isometries/__init__.py
@@ -1,0 +1,7 @@
+from .decompose_diagonal import _DecomposeDiagonal
+from .decompose_ucg import _DecomposeUCG
+from .single_qubit_gate import _SingleQubitGate
+from .decompose_isometry import _DecomposeIsometry
+from .apply_decompositions import (_apply_isometry,
+                                   _apply_diagonal_gate,
+                                   _apply_uniformly_controlled_gate)
