+++ projectq/isometries/decomposition.hpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include <array>
 #include <complex>
-#include <map>
+#include <omp.h>
 #include <cassert>
 #include <algorithm>
 #include <tuple>
