+++ projectq/ops/_diagonal_gate.py
@@ -11,15 +11,16 @@ class DiagonalGate(BasicGate):
     in the computational basis is diagonal.
 
     TODO:
-        D = DiagonalGate(angles)
+        D = DiagonalGate(complex_phases)
         D | qureg
 
     The order of the basis is given by the order of qureg....
     """
     def __init__(self, angles=[], phases=[]):
         # only ever need one of the two in any instance
-        assert not angles or not phases
+        assert len(angles) == 0 or len(phases) == 0
         if len(angles) > 0:
+            print("Dont construct from angles")
             self._angles = copy.copy(angles)
             self._phases = []
         elif len(phases) > 0:
@@ -28,10 +29,11 @@ class DiagonalGate(BasicGate):
         else:
             assert False
         self.interchangeable_qubit_indices = []
+        self._decomposed = False
 
     @property
     def angles(self):
-        if self._angles:
+        if len(self._angles) > 0:
             return self._angles
         else:
             print("not good 1")
@@ -39,15 +41,30 @@ class DiagonalGate(BasicGate):
 
     @property
     def phases(self):
-        if self._phases:
+        if len(self._phases) > 0:
             return self._phases
         else:
             print("not good 2")
             return [cmath.exp(1j*angle) for angle in self._angles]
 
     def get_inverse(self):
-        inv_angles = [-angle for angle in self.angles]
-        return DiagonalGate(angles = inv_angles)
+        inv_phases = [p.conjugate() for p in self.phases]
+        return DiagonalGate(phases = inv_phases)
+
+    def decompose(self):
+        assert self._decomposed == False
+        # don't use classes
+        from projectq.isometries import _DecomposeDiagonal
+        self._decomposition = _DecomposeDiagonal(self.phases).get_decomposition()
+        self._decomposed = True
+
+    @property
+    def decomposed(self):
+        return self._decomposed
+
+    @property
+    def decomposition(self):
+        return self._decomposition
 
     def __str__(self):
         return "D"
